Fix implicit cast warnings
This also fixes a bug introduced in b8f0fc40973b6ecb7818631db2f2fee888f1667e.

A 8 bit integer was being used in SocketTCP::accept() instead of a 32 bit one.

+++ include/mumble/Mumble.hpp
@@ -21,8 +21,8 @@ public:
 
 		uint32_t blob() const { return major << 16 | minor << 8 | patch; }
 
-		Version(const uint32_t blob) : major(blob >> 16), minor(blob >> 8), patch(blob) {}
-		Version(const uint16_t major, const uint16_t minor, const uint8_t patch)
+		Version(const uint32_t blob) : major(blob >> 16), minor(static_cast< uint8_t >(blob >> 8)), patch(static_cast< uint8_t >(blob)) {}
+		Version(const uint16_t major, const uint8_t minor, const uint8_t patch)
 			: major(major), minor(minor), patch(patch) {}
 	};
